Changes for key messages
Added a case for when every country has a key, when only 1 country has a key, and only when only a few number of countries have keys. The definition of "few number of countries" is based on the var cutoff. Also lists the countries with keys if lower than the var cutoff than the hundred+ countries without keys.